+++ ShopWithMeTests/ArtikelVerfuegbarkeitServiceTests.swift
@@ -10,7 +10,7 @@ struct ArtikelVerfuegbarkeitServiceTests {
             Artikel.self, ArtikelKategorie.self, Geschaeft.self, GeschaeftTyp.self,
             Einkaufsvorgang.self, KaufEintrag.self,
             Einkaufsliste.self, EinkaufslistenEintrag.self, SyncEvent.self,
-            ArtikelGeschaeftVerfuegbarkeit.self,
+            ArtikelGeschaeftVerfuegbarkeit.self, ArtikelListenKauf.self,
         ])
         let konfiguration = ModelConfiguration(schema: schema, isStoredInMemoryOnly: true)
         let container = try ModelContainer(for: schema, configurations: [konfiguration])